calcu/book/backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
openai
python-dotenv
markdown-it-py
requests
